// package/wagmiGenerated.ts
export const zoraCreator1155PreminterABI = [
  {
    type: 'error',
    inputs: [],
    name: 'INITIALIZABLE_CONTRACT_ALREADY_INITIALIZED',
  },
// ...
] as const